move insert product reposityor to service
